Moved CachableGroup to model.ts
