+++ src/redux/cache/index.ts
@@ -1,4 +1,4 @@
-export { CacheableGroup, AppCache, addOrReplaceCacheableGroup, findCacheableGroup } from './cache';
+export { AppCache, addOrReplaceCacheableGroup, findCacheableGroup } from './cache';
 export { AppCacheAction, AppCacheActionType, CacheGroupAction, AddToCacheAction } from './action';
 export { cacheGroup, addToCache } from './asyncActionCreator';
 export { appCacheReducer } from './reducer';
